Update: 衝撃検知の関数を別クラスに分ける
