+++ GpsMap/GpsMap_2/ViewControllerSearch.swift
@@ -198,70 +198,5 @@ extension ViewController: UISearchBarDelegate {
         }
         return MKOverlayRenderer()
     }
-    
-    // 転倒時の検出(加速度)
-    func fallDetectionAccel() -> Bool {
-       
-        // 加速度の合成値
-        self.mixAccel = sqrt(pow(self.xAccel,2)+pow(self.yAccel,2)+pow(self.zAccel,2))
-       
-        if self.mixAccel > contorlAccel + 0.1 || self.mixAccel < contorlAccel - 0.1 {
-            let flag = false
-            return flag
-        } else {
-            let flag = true
-            return flag
-        }
-    }
-    
-    // 転倒時の検出(ジャイロセンサ)
-    func fallDetectionGyro() -> Bool {
-        if self.roll > 60 || self.roll < -70   {
-         
-            if self.pitch < 30 &&  -30 < self.pitch {
-                
-                let flag = true
-                return flag
-                
-            } else if self.pitch > 150 {
-                
-                let flag = true
-                return flag
-                
-            } else if -180 < self.pitch && self.pitch < -150 {
-                
-                let flag = true
-                return flag
-                
-            } else {
-                
-                let flag = false
-                return flag
-                
-            }
- 
-        } else {
-            
-            let flag = false
-            return flag
-            
-        }
-    }
-    
-    // 転倒時の検出(気圧)
-    func fallDetectionPressure()-> Bool {
-        
-        if self.prePressure == 0 {
-            self.prePressure = self.pressure
-        }
-        
-        if fabs(self.pressure - self.prePressure) < 0.003 {
-            let flag = false
-            return flag
-        } else {
-            let flag = true
-            return flag
-        }
-    }
 
 }
